V7.3

+++ src/main/java/com/odk/Service/Interface/Service/ActiviteService.java
@@ -4,13 +4,17 @@ import com.odk.Entity.Activite;
 import com.odk.Entity.Utilisateur;
 import com.odk.Enum.Statut;
 import com.odk.Repository.ActiviteRepository;
+import com.odk.Repository.UtilisateurRepository;
 import com.odk.Service.Interface.CrudService;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.dao.DataAccessException;
+import org.springframework.http.HttpStatus;
 import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -24,10 +28,21 @@ public class ActiviteService implements CrudService<Activite, Long> {
     private PersonnelService personnelService;
     private EmailService emailService;
     private UtilisateurService utilisateurService;
+    private UtilisateurRepository utilisateurRepository;
 
     @Override
     public Activite add(Activite entity) {
         try {
+            // Récupérer l'utilisateur connecté
+            String email1 = SecurityContextHolder.getContext().getAuthentication().getName();
+            Utilisateur utilisateurPerso = utilisateurRepository.findByEmail(email1)
+                    .orElseThrow(() -> new RuntimeException("Utilisateur non trouvé"));
+
+            // Associer l'utilisateur comme créateur
+            entity.setCreatedBy(utilisateurPerso);
+
+            // Mettre à jour le statut de l'activité
+            entity.mettreAJourStatut();
             // Enregistrer l'activité
             Activite activiteCree = activiteRepository.save(entity);
 
@@ -36,7 +51,7 @@ public class ActiviteService implements CrudService<Activite, Long> {
 
              //  Filtrer les utilisateurs ayant le rôle "personnel"
             List<String> emailsPersonnel = utilisateurs.stream()
-                    .filter(utilisateur -> utilisateur.getRole().getNom().equals("Personnel")) // Vérifiez que le rôle est bien défini
+                    .filter(utilisateur -> utilisateur.getRole().getNom().equals("PERSONNEL")) // Vérifiez que le rôle est bien défini
                     .map(Utilisateur::getEmail) // Récupérer les emails
                     .collect(Collectors.toList());
 
@@ -58,6 +73,7 @@ public class ActiviteService implements CrudService<Activite, Long> {
             throw new RuntimeException("Erreur lors de la création de l'activité", e);
         }
     }
+
     @Override
     public List<Activite> List() {
         return activiteRepository.findAll();
@@ -83,7 +99,18 @@ public class ActiviteService implements CrudService<Activite, Long> {
     @Transactional
     @Override
     public Activite update(Activite activite, Long id) {
+        // Récupérer l'utilisateur connecté
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        Utilisateur utilisateur = utilisateurRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("Utilisateur non trouvé"));
+
         return activiteRepository.findById(id).map(a -> {
+            // Vérifier que l'utilisateur connecté est le créateur de l'activité
+            if (!a.getCreatedBy().getEmail().equals(utilisateur.getEmail())) {
+                throw new RuntimeException("Vous n'êtes pas autorisé à modifier cette activité.");
+            }
+
+            // Mettre à jour les champs modifiables
             if (activite.getNom() != null) {
                 a.setNom(activite.getNom());
             }
@@ -113,14 +140,32 @@ public class ActiviteService implements CrudService<Activite, Long> {
                 a.getEtape().addAll(activite.getEtape());
             }
 
+            // Mettre à jour le statut
+            a.mettreAJourStatut();
+
+            // Sauvegarder les modifications
             return activiteRepository.save(a);
-        }).orElseThrow(() -> new RuntimeException("Votre id n'existe pas"));
+        }).orElseThrow(() -> new RuntimeException("L'activité avec l'ID spécifié n'existe pas."));
     }
 
 
 
     @Override
     public void delete(Long id) {
+        // Récupérer l'utilisateur connecté
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        Utilisateur utilisateur = utilisateurRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("Utilisateur non trouvé"));
+
+        // Récupérer l'activité
+        Activite activite1 = activiteRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Activité non trouvée"));
+
+        // Vérifier si l'utilisateur est le créateur
+        if (!activite1.getCreatedBy().getId().equals(utilisateur.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Vous n'êtes pas autorisé à supprimer cette activité");
+        }
+
         Optional<Activite> activiteOptional = activiteRepository.findById(id);
         activiteOptional.ifPresent(activite -> activiteRepository.delete(activite));
     }
